Update rackuo test to kill the correct process at the end
Adding `2>&1` to the IO.popen command means that it opens an
additional process that wraps the rackup process, so the previous
version left rackup running at the end of the test run.

Using `-q` rather than `2>&1` means we capture the right process ID
to kill at the end of the test run.