Add basic validation of postcode, with error message
When a postcode is submitted to the find an assessor page
it checks that the submitted postcode is in a format that
matches our basic postcode regular expression.

If it doesn't then an error message is shown on the postcode
form field, and the original value is persisted in the
input field so it can be edited by the user.