Make default template variable behaviour to escape all
Based on the guidance on the sinatra FAQs[1], I have enabled
html escaping and set the default behaviour to be that all
template variables are escaped.

It is possible to "unescape" a template variable using the
following syntax:

```html
<%== html_safe_string %>
```

This should only be done on rare exceptions, and never to
user generated input (from citizen users, schemes, or assessors).

[1] http://sinatrarb.com/faq.html#auto_escape_html