Update bad RRN error message
"Enter a valid reference number" doesn't tell users everything
they need to know to be able to fix the error, and doesn't make
it clear that the input is malformed (an "invalid number" could
be one that doesn't exist).

The new message "Enter a 20-digit reference number" makes it
clearer what needs to be fixed for their search to succeed.