Only show "Recorded meter reading: " text on "meteringInstalled" note
The text "Recorded meter reading:" was being put in front of all notes
about various sections of the cooling plants section of air conditioning
reports. This has been reported to us as an error by an aircon software
provider.

This makes it so that the text is only added to the "meteringInstalled"
note, which is the one with a meter reading.