EPBR-497: Adding test if certificate returns 410
In this commit, I added a test to check if we return a 410 GONE error from the API if the certificate is CANCELLED or NOT_FOR_ISSUE. I didn't include it in the other test context as it technically exists (i.e. the cert with a valid assessment_id was lodged) but the assessment scheme revoked it.