Remove pinned bundler version
PaaS Ruby buildpack now has bundler 2.2.16.

This removes the pinned bundler version 2.1.4 in the build & test & deployment scripts.  This is for consistency with how we do bundler versioning in other repositories - we had a few ways of doing this, and as a team we've decided that it's better to use the latest supported version rather than pin to an old version.

Also replaces Gemfile.lock with one generated by the new bundler version.