Set EPB_UNLEASH_URI in all frontend app environments
Now toggles are integrated into the frontend the deployed
apps need to know where to find the Unleash server.

These values already exist in the parameter store as they
are already used by the API, so we just need to pick them
up and set them as part of application deployment, as here.