Reduce the gap under top EPC summary list
This reduces the space under the "blue box" on EPCs
by removing the bottom margin from the summary list
which generally has property type and floor area
information.  This space is where the "print this
certificate" text used to be.

The domestic EPC already had the "epc-box-container"
class, but needed to be added to the non-domestic EPC
component so behaviour will be the same for both
certificate types.