Update the back link and browser history after printing
When Print button is clicked we are adding the `print=true` to the query string
param. As a result the back link and back button in the browser takes us to the
certificate page in the online view. We want to override the url and the
browser history to take the user to the previous page (usually the search
results).

Co-authored-by: Barry <barry.halper@communities.gov.uk>