Display address line 3 in search results if line 2 is empty
We have some cases when a certificate is lodged with empty addressLine2 and
street name entered in the addressLine3. For those the street name was not
displayed on the search results page

Helper method .compact_address is used by both search results view and the
certificate view. We are passing all 4 address lines to this helper method in
the certificate view. We don't want to always do it on the results page as the
row would be too wide. For this reason decided to add conditional logic in the
list view, as opposed to modifying the helper method. The helper method deals
with addressLine2 being empty.