Filter latest certificate by createdAt if they were lodged on the same day
We have cases when two assessments for the same property were lodged on the
same day. Currently there is no way of determining which one is the latest,
most recent one as the XML only contains dates. We want to use the timestamp to
determine the latest assessment for the search results pages.

created_at column was added in September 2020 so this data is not available for
assessments lodged prior to that date.