Set cache control header on certificate pages
We have seen a lot of traffic recently with multiple
requests for the same certificate page within a short
time.  To reduce load on the API and frontend we can
cache certificate pages for a short time (1 minute).

Our CDN will use these headers to decide how long to
wait before refreshing the page.