Hide the "Paid off" text if no repaid_date exists
Some Green Deal measures have no "repaid_date" in their data.
We are currently showing the text "Paid off" next to these,
where there should be a date.

If that date is in the future then the page incorrectly
implies that the measure has been paid off, so we should
hide the "Paid off" text if there is no date set.