Make service names consistent in Welsh
The service names on the GOV.UK start pages were slightly different translations from those shown in the header in-service.
We have checked with the translator and they have confirmed that these are the better ones to keep:
  Find an Energy Certificate > Dod o hyd i Dystysgrif Ynni
  Get a new energy certificate > Cael tystysgrif ynni newydd

Also reduced some redundancy in the translations file (duplicated service name content).