Fix rackup test
I think the inclusion of "port-9393" was expected in the first line of output from Rack, but we have been using Puma for some time now and this doesn't appear in the fist line of Puma startup.  We don't really care what it sayys - just need to wait for the first line of output so we know the server has started (rackup tests fail if we don't wait for server to start).