Fix tests broken by newlines
This isn't ideal, as tests only broken for me locally as assertions fail because of newlines in expected text.
Other devs and CI pipeline seems happy without this addition, but it seems legit to add it :shrug:

Not going to put it everywhere unless we se this issue more widely than just this test.