Run frontend-build when building Docker image
For the frontend app we need to run the frontend-build task to make sure
sass is compiled and assets are copied to the right folders.

In epb-dev-tools we run this as part of a `docker compose` step when starting
a container.  However, for the new infrastructure it will be better to
build the frontend assets to include them in the image itself, rather
than load npm when containers start up in different environments (this
should ensure the exact same assets are in use in all environments using
the same Docker image).

This commit adds steps to the Dockerfile to :
1. install node
2. run the frontend-build task