Ensure locale is 'en' for unit tests that check translations
Since removing the Rack::Attack code (previous commit 0bf687dce582572316f0873cd97003e89a429bec) a number of unit tests fail because they unexpectedly get results in Welsh rather than the English they are testing for.

It seems that the locale persists between tests during the run, and any acceptance test that calls a URL for a page in Welsh toggles the shared locale to 'cy'. This is fine for acceptance tests because every time a URL is called by an acceptance test the correct locale will be determined from the URL.  However, for unit tests that don't exercise things via calls to routes the locale is just whatever it happened to be left as by the previous test.

It's not clear why this issue has only arisen now, but it seems that explicitly setting the locale for unit tests that expect translations in a particular language is a reasonable way to get these tests passing.