Add font correction for printed EPCs
This specifically tarhets the "type of assessment" row in the printed domestic EPC, which previously had "SAP" or "RdSAP" in a much larger font than other values in the table, due to being nested differently from other values (for the info drop-down).

This override in the print styles ensures the value is the same font size as other values in the printed certificate.