EPBR-6817 added co2 data to performance data csv
Refactors existing fetch stats use case and fetch cvs to inherit from the same base class. Moved the core functionality from the fetch_statistics_csv.rb to the existing domain class to better isolate tests. Created a new method of that class to create the object sent to controller to be converted into a csv. Added the key for the this class for the new data point.