Improve logic for show/hiding DEC info on DEC-RR
We have a small number of DEC-RR where the co-lodged DEC
has been cancelled but the RR remains in place.  In this
case the "relatedRrn" field exists in the /summary but
there is no "energyBandFromRelatedCertificate".

Previous logic hid the section from the page but not the
link from the page contents.

This updates the logic to show/hide the section to be the
same as to show/hide the contents link.

Also split out a test that previously confused related assessments
(i.e. previous DEC-RRs) with the "related_rrn" of a co-lodged
DEC.