Switched `quiet` flag for `silence_deprecations`.
The `quiet` flag was silencing useful deprecation warnings from govuk-frontend. We switched to the `silence_deprecations` flag instead so we can pass it specific deprecations we want silenced, related to known issues in govuk-frontend.

Previously we have used `quiet_deps` which doesn't work the way you'd expect due to how we load govuk-frontend.