Update conditions for rackup having started
Previously, our test code was waiting for just the first line of output after starting up the Puma webserver but (on some machines at least) this wasn't sufficient for the server to be ready for tests to commence.

By waiting for the final line of output on successful Puma startup we should hopefully avoid "Connection refused" errors for tests that attempt to connect before the server is ready.

The full output at server startup is currently:
    Bundler is using a binstub that was created for a different gem (rackup).
    You should run `bundle binstub rack` to work around a system/bundle conflict.
    Puma starting in single mode...
    * Puma version: 6.6.0 (\"Return to Forever\")
    * Ruby version: ruby 3.3.4 (2024-07-09 revision be1089c8ec) [arm64-darwin24]
    *  Min threads: 2
    *  Max threads: 4
    *  Environment: development
    *          PID: 5550
    * Listening on http://127.0.0.1:9393