Fix display of new RdSAP10.2 addenda
We missed adding the new addenda values to the frontend translation files, so EPCs with addenda 15-18 have "translation missing".  This adds the addenda titles and descriptions from the ExternalDefinitions file in the RdSAP10.2 schema.

Also adds conditional logic for addenda 13 and 14, which have changed their meanings in RdSAP 10.2 compared to historical schemas.