Update robots.txt
Counterintuitively, we need to allow the first page of the services to be crawled so that search engines can crawl and find the "noindex" meta tag and hopefully honour it.

The /find-a-certificate/type-of-property page is currently shown as the second search result in Google for "Find EPC", with "No information is available for this page." written below.  This change to the robots.txt should mean that the site is excluded altogether from search results.